♻️ refactor(c): LTO, split hot/cold headers, move query
Enable link-time optimization on the release wheel (pyproject cibuildwheel
config-settings) and the CodSpeed build (pgo_build.py), layered on top of PGO.
LTO reaches every compile command and the final link, verified with ninja -v.

With LTO on, split the hot CSS selector engine out of query/css/selector.h
into a compiled query/css/selector.c (types and entry prototypes stay in the
header); LTO re-inlines the matcher across the new unit boundary, which the
spike for #478 measured recovers the split to within 0.1% of the monolith.
Move the cold CSS-to-XPath translator to query/css/to_xpath.c, and lift the
Node select, regex, and XPath query bindings out of dom/element.c into
query/methods.c, shrinking element.c by ~1600 lines. handle_drop_index and
attr_key_utf8 stay in element.c and gain declarations in nodes.h.

tree.c now includes its foreign-content and quirks helpers at the top instead
of mid-file; foreign.h forward-declares the tree-builder statics it shares and
both headers gain include guards.

Parser and query output are byte-identical; 100% line and branch coverage holds
on both llvm-cov and gcc. Part of #478.
