+++ meson.build
@@ -47,6 +47,9 @@ py.extension_module(
         'src/turbohtml/_c/serialize/js/mangle.c',
         'src/turbohtml/_c/serialize/js/minify.c',
         'src/turbohtml/_c/serialize/js/lexdump.c',
+        'src/turbohtml/_c/query/css/selector.c',
+        'src/turbohtml/_c/query/css/to_xpath.c',
+        'src/turbohtml/_c/query/methods.c',
         'src/turbohtml/_c/query/xpath/ast.c',
         'src/turbohtml/_c/query/xpath/lexer.c',
         'src/turbohtml/_c/query/xpath/parser.c',
